Revert "Added a composable waitForCondition Driver/extension API (#37736)" (#38815)
This reverts commit 8edc3f76cc53375a18057b1cd51eb7c2a85a2769.
