+++ packages/flutter_driver/lib/src/common/find.dart
@@ -106,6 +106,52 @@ class WaitForAbsentResult extends Result {
   Map<String, dynamic> toJson() => <String, dynamic>{};
 }
 
+/// A Flutter Driver command that waits until there are no more transient callbacks in the queue.
+class WaitUntilNoTransientCallbacks extends Command {
+  /// Creates a command that waits for there to be no transient callbacks.
+  const WaitUntilNoTransientCallbacks({ Duration timeout }) : super(timeout: timeout);
+
+  /// Deserializes this command from the value generated by [serialize].
+  WaitUntilNoTransientCallbacks.deserialize(Map<String, String> json)
+    : super.deserialize(json);
+
+  @override
+  String get kind => 'waitUntilNoTransientCallbacks';
+}
+
+/// A Flutter Driver command that waits until the frame is synced.
+class WaitUntilNoPendingFrame extends Command {
+  /// Creates a command that waits until there's no pending frame scheduled.
+  const WaitUntilNoPendingFrame({ Duration timeout }) : super(timeout: timeout);
+
+  /// Deserializes this command from the value generated by [serialize].
+  WaitUntilNoPendingFrame.deserialize(Map<String, String> json)
+    : super.deserialize(json);
+
+  @override
+  String get kind => 'waitUntilNoPendingFrame';
+}
+
+/// A Flutter Driver command that waits until the Flutter engine rasterizes the
+/// first frame.
+///
+/// {@template flutter.frame_rasterized_vs_presented}
+/// Usually, the time that a frame is rasterized is very close to the time that
+/// it gets presented on the display. Specifically, rasterization is the last
+/// expensive phase of a frame that's still in Flutter's control.
+/// {@endtemplate}
+class WaitUntilFirstFrameRasterized extends Command {
+  /// Creates this command.
+  const WaitUntilFirstFrameRasterized({ Duration timeout }) : super(timeout: timeout);
+
+  /// Deserializes this command from the value generated by [serialize].
+  WaitUntilFirstFrameRasterized.deserialize(Map<String, String> json)
+      : super.deserialize(json);
+
+  @override
+  String get kind => 'waitUntilFirstFrameRasterized';
+}
+
 /// Base class for Flutter Driver finders, objects that describe how the driver
 /// should search for elements.
 abstract class SerializableFinder {
