+++ packages/flutter_driver/test/flutter_driver_test.dart
@@ -6,7 +6,6 @@ import 'dart:async';
 
 import 'package:flutter_driver/src/common/error.dart';
 import 'package:flutter_driver/src/common/health.dart';
-import 'package:flutter_driver/src/common/wait.dart';
 import 'package:flutter_driver/src/driver/driver.dart';
 import 'package:flutter_driver/src/driver/timeline.dart';
 import 'package:json_rpc_2/json_rpc_2.dart' as rpc;
@@ -249,41 +248,12 @@ void main() {
       });
     });
 
-    group('waitForCondition', () {
-      test('sends the wait for NoPendingFrameCondition command', () async {
-        when(mockIsolate.invokeExtension(any, any)).thenAnswer((Invocation i) {
-          expect(i.positionalArguments[1], <String, dynamic>{
-            'command': 'waitForCondition',
-            'timeout': _kSerializedTestTimeout,
-            'conditionName': 'NoPendingFrameCondition',
-          });
-          return makeMockResponse(<String, dynamic>{});
-        });
-        await driver.waitForCondition(const NoPendingFrameCondition(), timeout: _kTestTimeout);
-      });
-
-      test('sends the waitForCondition of combined conditions command', () async {
-        when(mockIsolate.invokeExtension(any, any)).thenAnswer((Invocation i) {
-          expect(i.positionalArguments[1], <String, dynamic>{
-            'command': 'waitForCondition',
-            'timeout': _kSerializedTestTimeout,
-            'conditionName': 'CombinedCondition',
-            'conditions': '[{"conditionName":"NoPendingFrameCondition"},{"conditionName":"NoTransientCallbacksCondition"}]',
-          });
-          return makeMockResponse(<String, dynamic>{});
-        });
-        const WaitCondition combinedCondition = CombinedCondition(<WaitCondition>[NoPendingFrameCondition(), NoTransientCallbacksCondition()]);
-        await driver.waitForCondition(combinedCondition, timeout: _kTestTimeout);
-      });
-    });
-
     group('waitUntilNoTransientCallbacks', () {
       test('sends the waitUntilNoTransientCallbacks command', () async {
         when(mockIsolate.invokeExtension(any, any)).thenAnswer((Invocation i) {
           expect(i.positionalArguments[1], <String, dynamic>{
-            'command': 'waitForCondition',
+            'command': 'waitUntilNoTransientCallbacks',
             'timeout': _kSerializedTestTimeout,
-            'conditionName': 'NoTransientCallbacksCondition',
           });
           return makeMockResponse(<String, dynamic>{});
         });
