+++ packages/flutter_driver/test/src/extension_test.dart
@@ -12,7 +12,6 @@ import 'package:flutter_driver/src/common/find.dart';
 import 'package:flutter_driver/src/common/geometry.dart';
 import 'package:flutter_driver/src/common/request_data.dart';
 import 'package:flutter_driver/src/common/text.dart';
-import 'package:flutter_driver/src/common/wait.dart';
 import 'package:flutter_driver/src/extension/extension.dart';
 import 'package:flutter_test/flutter_test.dart';
 
@@ -29,18 +28,18 @@ void main() {
     });
 
     testWidgets('returns immediately when transient callback queue is empty', (WidgetTester tester) async {
-      extension.call(const WaitUntilNoTransientCallbacks().serialize()) // ignore: deprecated_member_use_from_same_package
-          .then<void>(expectAsync1((Map<String, dynamic> r) {
-        result = r;
-      }));
+      extension.call(const WaitUntilNoTransientCallbacks().serialize())
+        .then<void>(expectAsync1((Map<String, dynamic> r) {
+          result = r;
+        }));
 
       await tester.idle();
       expect(
-        result,
-        <String, dynamic>{
-          'isError': false,
-          'response': null,
-        },
+          result,
+          <String, dynamic>{
+            'isError': false,
+            'response': null,
+          },
       );
     });
 
@@ -49,10 +48,10 @@ void main() {
         // Intentionally blank. We only care about existence of a callback.
       });
 
-      extension.call(const WaitUntilNoTransientCallbacks().serialize()) // ignore: deprecated_member_use_from_same_package
-          .then<void>(expectAsync1((Map<String, dynamic> r) {
-        result = r;
-      }));
+      extension.call(const WaitUntilNoTransientCallbacks().serialize())
+        .then<void>(expectAsync1((Map<String, dynamic> r) {
+          result = r;
+        }));
 
       // Nothing should happen until the next frame.
       await tester.idle();
@@ -61,11 +60,11 @@ void main() {
       // NOW we should receive the result.
       await tester.pump();
       expect(
-        result,
-        <String, dynamic>{
-          'isError': false,
-          'response': null,
-        },
+          result,
+          <String, dynamic>{
+            'isError': false,
+            'response': null,
+          },
       );
     });
 
@@ -74,173 +73,6 @@ void main() {
       final dynamic result = RequestDataResult.fromJson((await extension.call(const RequestData('hello').serialize()))['response']);
       expect(log, <String>['hello']);
       expect(result.message, '1');
-    });
-  });
-
-  group('waitForCondition', () {
-    FlutterDriverExtension extension;
-    Map<String, dynamic> result;
-    int messageId = 0;
-    final List<String> log = <String>[];
-
-    setUp(() {
-      result = null;
-      extension = FlutterDriverExtension((String message) async { log.add(message); return (messageId += 1).toString(); }, false);
-    });
-
-    testWidgets('waiting for NoTransientCallbacksCondition returns immediately when transient callback queue is empty', (WidgetTester tester) async {
-      extension.call(const WaitForCondition(NoTransientCallbacksCondition()).serialize())
-          .then<void>(expectAsync1((Map<String, dynamic> r) {
-        result = r;
-      }));
-
-      await tester.idle();
-      expect(
-        result,
-        <String, dynamic>{
-          'isError': false,
-          'response': null,
-        },
-      );
-    });
-
-    testWidgets('waiting for NoTransientCallbacksCondition returns until no transient callbacks', (WidgetTester tester) async {
-      SchedulerBinding.instance.scheduleFrameCallback((_) {
-        // Intentionally blank. We only care about existence of a callback.
-      });
-
-      extension.call(const WaitForCondition(NoTransientCallbacksCondition()).serialize())
-          .then<void>(expectAsync1((Map<String, dynamic> r) {
-        result = r;
-      }));
-
-      // Nothing should happen until the next frame.
-      await tester.idle();
-      expect(result, isNull);
-
-      // NOW we should receive the result.
-      await tester.pump();
-      expect(
-        result,
-        <String, dynamic>{
-          'isError': false,
-          'response': null,
-        },
-      );
-    });
-
-    testWidgets('waiting for NoPendingFrameCondition returns immediately when frame is synced', (
-        WidgetTester tester) async {
-      extension.call(const WaitForCondition(NoPendingFrameCondition()).serialize())
-          .then<void>(expectAsync1((Map<String, dynamic> r) {
-        result = r;
-      }));
-
-      await tester.idle();
-      expect(
-        result,
-        <String, dynamic>{
-          'isError': false,
-          'response': null,
-        },
-      );
-    });
-
-    testWidgets('waiting for NoPendingFrameCondition returns until no pending scheduled frame', (WidgetTester tester) async {
-      SchedulerBinding.instance.scheduleFrame();
-
-      extension.call(const WaitForCondition(NoPendingFrameCondition()).serialize())
-          .then<void>(expectAsync1((Map<String, dynamic> r) {
-        result = r;
-      }));
-
-      // Nothing should happen until the next frame.
-      await tester.idle();
-      expect(result, isNull);
-
-      // NOW we should receive the result.
-      await tester.pump();
-      expect(
-        result,
-        <String, dynamic>{
-          'isError': false,
-          'response': null,
-        },
-      );
-    });
-
-    testWidgets(
-        'waiting for combined conditions returns immediately', (WidgetTester tester) async {
-      const WaitCondition combinedCondition = CombinedCondition(<WaitCondition>[NoTransientCallbacksCondition(), NoPendingFrameCondition()]);
-      extension.call(const WaitForCondition(combinedCondition).serialize())
-          .then<void>(expectAsync1((Map<String, dynamic> r) {
-        result = r;
-      }));
-
-      await tester.idle();
-      expect(
-        result,
-        <String, dynamic>{
-          'isError': false,
-          'response': null,
-        },
-      );
-    });
-
-    testWidgets(
-        'waiting for combined conditions returns until no transient callbacks', (WidgetTester tester) async {
-      SchedulerBinding.instance.scheduleFrame();
-      SchedulerBinding.instance.scheduleFrameCallback((_) {
-        // Intentionally blank. We only care about existence of a callback.
-      });
-
-      const WaitCondition combinedCondition = CombinedCondition(<WaitCondition>[NoTransientCallbacksCondition(), NoPendingFrameCondition()]);
-      extension.call(const WaitForCondition(combinedCondition).serialize())
-          .then<void>(expectAsync1((Map<String, dynamic> r) {
-        result = r;
-      }));
-
-      // Nothing should happen until the next frame.
-      await tester.idle();
-      expect(result, isNull);
-
-      // NOW we should receive the result.
-      await tester.pump();
-      expect(
-        result,
-        <String, dynamic>{
-          'isError': false,
-          'response': null,
-        },
-      );
-    });
-
-    testWidgets(
-        'waiting for combined conditions returns until no pending scheduled frame', (WidgetTester tester) async {
-      SchedulerBinding.instance.scheduleFrame();
-      SchedulerBinding.instance.scheduleFrameCallback((_) {
-        // Intentionally blank. We only care about existence of a callback.
-      });
-
-      const WaitCondition combinedCondition = CombinedCondition(<WaitCondition>[NoPendingFrameCondition(), NoTransientCallbacksCondition()]);
-      extension.call(const WaitForCondition(combinedCondition).serialize())
-          .then<void>(expectAsync1((Map<String, dynamic> r) {
-        result = r;
-      }));
-
-      // Nothing should happen until the next frame.
-      await tester.idle();
-      expect(result, isNull);
-
-      // NOW we should receive the result.
-      await tester.pump();
-      expect(
-        result,
-        <String, dynamic>{
-          'isError': false,
-          'response': null,
-        },
-      );
     });
   });
 
@@ -513,7 +345,7 @@ void main() {
 
     testWidgets('returns immediately when frame is synced', (
         WidgetTester tester) async {
-      extension.call(const WaitUntilNoPendingFrame().serialize()) // ignore: deprecated_member_use_from_same_package
+      extension.call(const WaitUntilNoPendingFrame().serialize())
           .then<void>(expectAsync1((Map<String, dynamic> r) {
         result = r;
       }));
@@ -534,7 +366,7 @@ void main() {
         // Intentionally blank. We only care about existence of a callback.
       });
 
-      extension.call(const WaitUntilNoPendingFrame().serialize()) // ignore: deprecated_member_use_from_same_package
+      extension.call(const WaitUntilNoPendingFrame().serialize())
           .then<void>(expectAsync1((Map<String, dynamic> r) {
         result = r;
       }));
@@ -558,7 +390,7 @@ void main() {
         'waits until no pending scheduled frame', (WidgetTester tester) async {
       SchedulerBinding.instance.scheduleFrame();
 
-      extension.call(const WaitUntilNoPendingFrame().serialize()) // ignore: deprecated_member_use_from_same_package
+      extension.call(const WaitUntilNoPendingFrame().serialize())
           .then<void>(expectAsync1((Map<String, dynamic> r) {
         result = r;
       }));
